+++ packages/create-farm-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-farm-plugin
 
+## 0.1.14
+
+### Patch Changes
+
+- 3adeb1a8: Update Cargo.lock of rust plugin
+
 ## 0.1.13
 
 ### Patch Changes